api authentication done using jwt

+++ controllers/api/v1/posts_api.js
@@ -18,8 +18,9 @@ module.exports.index= async function(req,res){
 module.exports.destroy = async function(req,res){
     try{
         let post= await Post.findById(req.params.id);
-
-        // if(post.user == req.user.id){
+        console.log("Hello Guys");
+        
+        if(post.user == req.user.id){
             post.remove();
             
             await Comment.deleteMany({post:req.params.id});
@@ -28,8 +29,13 @@ module.exports.destroy = async function(req,res){
             return res.status(200).json({
                 message: 'Deleted successfully'
             });
+        }else{
+            return res.status(401).json({
+                message: 'You cannot delete this post'
+            }); 
+        }
         
-    
+     
         
     }
     catch(err)
